sound: Add busy wait to write port status on DSP (#88)
* Updated project settings OPTIMIZATION_MODE and FITTER_AGGRESSIVE_ROUTABILITY_OPTIMIZATION so the design fits.

* Updated the Sound DSP module to look busy while playing sound and receiving DMA commands. Allowed new sounds to play immediately if DMA is paused.

Games, such as The Secret to Monkey Island, have a compiled in CT-VOICE driver
that communicates to the DSP. When the driver knows a sound is currently playing,
it adds an additional step before starting to play the new sound. The driver will poll
the write status port (2xCh) until it sees a busy flag, and if it sees it then it sends
the pause command first. If the driver never sees a busy status it will eventually
timeout polling and skip sending the pause command.

Without setting the busy flag a pause command is not sent, the game becomes out
of sync with the number of bytes it expects to transfer and locks up.

This change was confirmed to fix freezes during The Secret of Monkey Island,
Monkey Island 2, and Jill of the Jungle.

* Fixed spacing.